Extract and use primite renderers
