+++ packages/charts/src/chart_types/xy_chart/renderer/canvas/axes/tick.ts
@@ -8,11 +8,11 @@
 
 import { AxisProps } from '.';
 import { colorToRgba, RgbaTuple } from '../../../../../common/color_library_wrappers';
+import { renderMultiLine } from '../../../../../renderers/canvas/primitives/line';
 import { Position } from '../../../../../utils/common';
 import { isHorizontalAxis } from '../../../utils/axis_type_utils';
 import { AxisTick } from '../../../utils/axis_utils';
 import { HIDE_MINOR_TIME_GRID, HIERARCHICAL_GRID_WIDTH, OUTSIDE_RANGE_TOLERANCE } from '../../../utils/grid_lines';
-import { renderMultiLine } from '../primitives/line';
 
 const BASELINE_CORRECTION = 2; // the bottom of the em is a bit higher than the bottom alignment; todo consider measuring
