+++ packages/charts/src/chart_types/xy_chart/renderer/canvas/primitives/path.ts
@@ -6,10 +6,10 @@
  * Side Public License, v 1.
  */
 
-import { renderMultiLine } from './line';
 import { RGBATupleToString } from '../../../../../common/color_library_wrappers';
 import { Fill, Rect, Stroke } from '../../../../../geoms/types';
 import { withClipRanges } from '../../../../../renderers/canvas';
+import { renderMultiLine } from '../../../../../renderers/canvas/primitives/line';
 import { ClippedRanges } from '../../../../../utils/geometry';
 import { Point } from '../../../../../utils/point';
