+++ packages/charts/src/chart_types/xy_chart/renderer/canvas/primitives/shapes.ts
@@ -6,10 +6,10 @@
  * Side Public License, v 1.
  */
 
-import { MIN_STROKE_WIDTH } from './line';
 import { RGBATupleToString } from '../../../../../common/color_library_wrappers';
 import { Circle, Fill, Stroke } from '../../../../../geoms/types';
 import { withContext } from '../../../../../renderers/canvas';
+import { MIN_STROKE_WIDTH } from '../../../../../renderers/canvas/primitives/line';
 import { degToRad } from '../../../../../utils/common';
 import { PointShape } from '../../../../../utils/themes/theme';
 import { ShapeRendererFn } from '../../shapes_paths';
