+++ packages/charts/src/renderers/canvas/primitives/text.ts
@@ -14,9 +14,6 @@ import { degToRad } from '../../../utils/common';
 import { Point } from '../../../utils/point';
 
 /** @internal */
-export type TextDirection = 'rtl' | 'ltr'; // TODO: export this
-
-/** @internal */
 export type TextFont = Font & {
   fontSize: number;
   align: TextAlign;
